+++ projects/amdsmi/tests/amd_smi_test/functional/memory_read_write.cc
@@ -235,8 +235,8 @@ void TestMemoryReadWrite::Run(void) {
 
       // Convert to GB for display
       const long page_size = sysconf(_SC_PAGESIZE);
-      double gb =
-          (static_cast<double>(ttm_info.current_pages) * page_size) / (1024.0 * 1024.0 * 1024.0);
+      double gb = (static_cast<double>(ttm_info.current_pages) * static_cast<double>(page_size)) /
+                  (1024.0 * 1024.0 * 1024.0);
       std::cout << "\t**Current TTM Size: " << gb << " GB" << std::endl;
     }
